+++ src/components/ContentHeader.jsx
@@ -12,7 +12,7 @@ function ContentHeader({
   ...props
 }) {
   return (
-    <header className="pb-4 w-full max-w-full-sidebar fixed bg-white h-24">
+    <header className="pb-4 w-full max-w-full-sidebar fixed bg-white h-24 border-b">
       <div className="flex justify-between items-end h-full">
         {withHistory && <BackButton />}
         <h2 className="font-bold text-lg sm:text-xl md:text-3xl">
